// src/modules/auth/auth.schemas.ts
import { UserRole } from '@prisma/client';
import { z } from 'zod';

export const registerSchema = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  password: z.string().min(8),
  role: z.nativeEnum(UserRole).default(UserRole.CANDIDATE),
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8),
});